Cpu+Debugger: separate the debugger to a separate struct
This way its easier to manage, and I'm planning to have the debugger
enabled by a feature, which will allow to easily disable it if we don't
need it.
